Gave in and used SIGKILL on children and then self
This seems like a horrible idea, but since the children are almost
ALWAYS in a blocked state, they ignore calls to SIGTERM, and then
continue blocking while the master runs a piece of code in the
_bootstrap method of multiprocessing.Process class that ultimately
calls waitpid on each child. Perhaps they are blocking on a resource
the master holds, but I have set everything in child not to block at all,
except for the blocking that occurs in urllib.urlopen, which can't
be avoided.

In retrospect, since the children are almost certainly IO bound, it probably
would have been ok to just use the Python threading module. The Global Interpreter
Lock is only a concern on CPU-bound processes, apparently.

I will return to this, but for the time being, I am going to start planning
database insertions, and try to work on a graceful exit for non KeyboardInterrupt
situations, for instance when a certain amount of data has been inserted into the db.