It seriously seems like everything is working fine without catching KeyboardInterrupt in the master process, or when I just reraise the exception, or when I catch it and just pass.
Last time when I tried this, I defined an atexit function. I'm guessing that
that function was screwing with the way multiprocessing exits. Not totally sure,
but I am going to proceed cautiously using this scheme and revert if it turns out
that a problem still exists. Things could get tricky when I try to implement
a "nice" exit strategy.