Resolved some nagging URL parsing issues
I decided that I'm not going to get caught in the details on this one.
The scheme now goes like this:

urljoin first to see if relative link can be fixed with base_url
then when we get to hostname, if it doesn't have 3 parts (a.k.a.
subdomain, domain, and suffix), just give up.

This refactor also ensures that only pristine urls are placed onto
the work queue, which simplifies everything.